test(enrichment): split media enricher tests into per-provider modules
