+++ backend/tests/enrichment/__init__.py
@@ -1,0 +1,1 @@
+"""Enrichment test modules."""
